Students/01_SecondWeek/move/ME403_LabFiles_Yunus/magician/requirements.txt
pydobotplus
pyserial
numpy
